requirements.txt
aniso8601==7.0.0
asgiref==3.8.1
astroid==3.1.0
autopep8==2.1.0
dill==0.3.8
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
isort==5.13.2
mccabe==0.7.0
pep8==1.7.1
platformdirs==4.2.1
promise==2.3
pycodestyle==2.11.1
PyJWT==1.7.1
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tomlkit==0.12.4
